loading/lazyload/original-base-url-applied-iframe-tentative.html

<!DOCTYPE html>
<head>
  <title>Deferred iframes with loading='lazy' use the original
         base URL specified at the parse time</title>
  <link rel="author" title="Rob Buis" href="mailto:rbuis@igalia.com">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
  <base href='/loading/lazyload/resources/'>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<script>
  let has_window_loaded = false;

  async_test(function(t) {
    const below_viewport_iframe = new ElementLoadPromise("below_viewport_iframe");

    // Change the base URL and scroll down to load the deferred elements.
    window.addEventListener("load", () => {
      var base = document.getElementsByTagName('base')[0];
      base.href = '/invalid-url-where-no-subresources-exist/';
      has_window_loaded = true;
      below_viewport_iframe.element().scrollIntoView();
    });

    below_viewport_iframe.promise.then(
      t.step_func_done(function() {
        assert_true(has_window_loaded);
        assert_true(below_viewport_iframe.element().contentDocument.body.
          innerHTML.includes("<p>Subframe</p>"));
      }));
  }, "Test that when deferred iframe is loaded, the parse time base URL is used.");
</script>

<body>
  <div style="height:110vh"></div>
  <iframe id="below_viewport_iframe" src="subframe.html" loading="lazy"
    width="200px" height="100px" onload="below_viewport_iframe.resolve();">
  </iframe>
</body>
